+++ ipa_room_exploration/ros/include/ipa_room_exploration/room_exploration_action_server.h
@@ -9,6 +9,7 @@
 #include <dynamic_reconfigure/server.h>
 // OpenCV specific
 #include <opencv/cv.h>
+#include <opencv/highgui.h>
 // Eigen library
 #include <Eigen/Dense>
 // standard c++ libraries
@@ -27,6 +28,8 @@
 #include <geometry_msgs/Polygon.h>
 #include <geometry_msgs/Point32.h>
 #include <nav_msgs/OccupancyGrid.h>
+#include <sensor_msgs/Image.h>
+#include <sensor_msgs/image_encodings.h>
 // specific from this package
 #include <ipa_room_exploration/concorde_TSP.h>
 #include <ipa_room_exploration/RoomExplorationConfig.h>
@@ -193,17 +196,6 @@ protected:
 		cv::warpAffine(map, dst, rot_mat, map.size());
 		cv::flip(dst, map, 1);
 	}
-
-	// function to draw the points that have been covered by the field of view, when the robot moved trough the room
-	//		--> use given Poses and original field of view points to do so
-	void drawSeenPoints(cv::Mat& reachable_areas_map, const std::vector<geometry_msgs::Pose2D>& robot_poses,
-			const std::vector<geometry_msgs::Point32>& field_of_view_points, const Eigen::Matrix<float, 2, 1> raycasting_corner_1,
-			const Eigen::Matrix<float, 2, 1> raycasting_corner_2, const float map_resolution, const cv::Point2d map_origin);
-
-	// function to draw the robot footprint into the given map, when coverage for the robot is wanted
-	void drawSeenPoints(cv::Mat& reachable_areas_map, const std::vector<geometry_msgs::Pose2D>& robot_poses,
-				const std::vector<geometry_msgs::Point32>& robot_footprint, const float map_resolution,
-				const cv::Point2d map_origin);
 
 	// !!Important!!
 	//  define the Nodehandle before the action server, or else the server won't start