+++ ipa_room_exploration/CHANGELOG.rst
@@ -1,6 +1,10 @@
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-Changelog for package ipa_building_msgs
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+Changelog for package ipa_room_exploration
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+0.1.1 (2017-02-10)
+------------------
+* excluded coverage check to standalone service server
+* Contributors: Florian Jordan
 0.1.0 (2017-02-09)
 ------------------
 * first completely working version of package